// src/escpos.types.d.ts
export type FEED_CONTROL_TYPE = 'LF' | 'GLF' | 'FF' | 'CR' | 'HT' | 'VT';
export type BITMAP_FORMAT_TYPE = 'S8' | 'D8' | 'S24' | 'D24';
export type QRCODE_LEVEL = 'L' | 'M' | 'Q' | 'H';

/**
 * `B` = Bold
 * `I` = Italic
 * `U` = Underline
 */
export type TXT_STYLE =
  | 'NORMAL'
  | 'B'
  | 'I'
  | 'U'
  | 'U2'
  | 'BI'
  | 'BIU'
  | 'BIU2'
  | 'BU'
  | 'BU2'
  | 'IU'
  | 'IU2';

export type MIME_TYPE =
  | 'image/png'
  | 'image/jpg'
  | 'image/jpeg'
  | 'image/gif'
  | 'image/bmp';
export type BARCODE_TYPE =
  | 'UPC_A'
  | 'UPC_E'
  | 'EAN13'
  | 'EAN8'
  | 'CODE39'
  | 'ITF'
  | 'NW7'
  | 'CODE93'
  | 'CODE128';

/**
 * `LT` = Left
 * `CT` = Center
 * `RT` = Right
 */
export type TXT_ALIGN = 'LT' | 'CT' | 'RT';
// ...
export interface Adapter {
  open(...args: any[]): Adapter;

  write(data: Buffer, callback: (error?: any) => void): Adapter;
}
// ...
export class Network implements Adapter {
  constructor(address: string, port?: number);

  open(callback?: (error: any, device: Network) => void): Network;

  close(callback?: (error: any, device: Network) => void): Network;

  write(data: Buffer, callback: (error?: any) => void): Network;
}

export class Console implements Adapter {
  constructor(handler?: (data: any[][]) => void);

  open(callback?: (error?: any) => void): Console;

  write(data: Buffer): Console;
}

export class Image {
  constructor(pixels: any);

  static load(url: string, callback?: (result: Image | Error) => void): void;
  static load(
    url: string,
    type: MIME_TYPE,
    callback?: (result: Image | Error) => void,
  ): void;

  size(): { width: number; height: number; colors: number };

  toBitmap(density?: number): {
    data: any;
    density: number;
  };

  toRaster(): {
    data: number[];
    width: number;
    height: number;
  };
}

export class Printer {
  constructor(adapter: Adapter, options?: { encoding?: string | undefined });

  static create(device: Adapter): Promise<Printer>;

  /**
   * Set printer model to recognize model-specific commands.
   * Supported models: [ null, 'qsprinter' ]
   *
   * For generic printers, set model to null
   */
  model(_model: string): Printer;

  marginBottom(size: number): Printer;

  marginLeft(size: number): Printer;

  marginRight(size: number): Printer;

  /**
   * Print raw
   */
  print(content: any): Printer;

  /**
   * Print raw with EOL
   */
  println(content: any): Printer;

  newLine(): Printer;

  /**
   * Print text with EOL
   */
  text(content: string, encoding?: string): Printer;

  /**
   * Draw a horizontal line with EOL
   *
   * Example:
   * `-------` 48x
   */
  drawLine(): Printer;

  /**
   * Print list of string with EOL
   */
  table(data: string[], encoding?: string): Printer;

  tableCustom(
    data:
      | {
          text: string;
          width: number;
          align: 'LEFT' | 'CENTER' | 'RIGHT';
        }
      | {
          text: string;
          cols: number;
          align: 'LEFT' | 'CENTER' | 'RIGHT';
        },
    encoding?: string,
  ): Printer;

  /**
   * Print text
   */
  pureText(content: string, encoding?: string): Printer;

  /**
   * Set default encoding
   * @default `GB18030` (Chinese)
   */
  encode(encoding: string): Printer;

  /**
   * Print blank lines
   */
  feed(n?: number): Printer;

  /**
   * Carrier feed and tabs.
   */
  control(ctrl: FEED_CONTROL_TYPE): Printer;

  align(align: TXT_ALIGN): Printer;

  /**
   * @default 'A'
   */
  font(family: 'A' | 'B'): Printer;

  style(type: TXT_STYLE): Printer;

  /**
   * Set text size
   *
   * 1 is for regular size, and 2 is twice the standard size.
   *
   * @default `1`
   */
  size(width: 1 | 2, height: 1 | 2): Printer;

  /**
   * Set character spacing
   */
  spacing(n?: number): Printer;

  /**
   * Set line spacing
   */
  lineSpace(n?: number): Printer;

  hardware(hw: string): Printer;

  barcode(
    code: string,
    type: BARCODE_TYPE,
    options?: {
      width: number;
      height: number;
      /**
       * OFF, ABOVE, BELOW, BOTH
       *
       * @default BELOW
       */
      position: 'OFF' | 'ABV' | 'BLW' | 'BTH';
      font: 'A' | 'B';
      includeParity: boolean;
    },
  ): Printer;

  qrcode(
    code: string,
    version: number,
    level: QRCODE_LEVEL,
    size: number,
  ): Printer;

  qrimage(
    content: string,
    callback?: (error: Error | null, printer?: Printer) => void,
  ): Printer;
  qrimage(
    content: string,
    options?: { type: string; mode: string },
    callback?: (error: Error | null, printer?: Printer) => void,
  ): Printer;

  async image(image: Image, density?: BITMAP_FORMAT_TYPE): Promise<Printer>;

  raster(image: Image, mode?: string): Printer;

  /**
   * Send pulse to kick the cash drawer
   *
   * @default 2
   */
  cashdraw(pin?: number): Printer;

  /**
   * Printer Buzzer (Beep sound)
   */
  beep(numberOfBuzzer: number, time: number): Printer;

  /**
   * Send data to hardware and flush buffer
   */
  flush(callback?: (error?: any) => void): Printer;

  /**
   * Cut paper.
   *
   * *** Cut will flush buffer to printer ***
   *
   * @param part If true, PAPER_PART_CUT, else PAPER_FULL_CUT
   * @param feed
   */
  cut(part?: boolean, feed?: number): Printer;

  close(callback?: (error?: any) => void, options?: any): Printer;

  /**
   * Select between two print color modes, if your printer supports it
   * @param color - 0 for primary color (black) 1 for secondary color (red)
   */
  color(color: 0 | 1): Printer;
}